test(eighth): add rooms, attendance and signup tests
