+++ Ion.egg-info/SOURCES.txt
@@ -82,6 +82,7 @@ docs/sourcedoc/intranet.apps.eighth.forms.rst
 docs/sourcedoc/intranet.apps.eighth.management.commands.rst
 docs/sourcedoc/intranet.apps.eighth.management.rst
 docs/sourcedoc/intranet.apps.eighth.rst
+docs/sourcedoc/intranet.apps.eighth.tests.rst
 docs/sourcedoc/intranet.apps.eighth.views.admin.rst
 docs/sourcedoc/intranet.apps.eighth.views.rst
 docs/sourcedoc/intranet.apps.emailfwd.rst
@@ -231,7 +232,6 @@ intranet/apps/eighth/models.py
 intranet/apps/eighth/notifications.py
 intranet/apps/eighth/serializers.py
 intranet/apps/eighth/tasks.py
-intranet/apps/eighth/tests.py
 intranet/apps/eighth/urls.py
 intranet/apps/eighth/utils.py
 intranet/apps/eighth/forms/__init__.py
@@ -322,6 +322,10 @@ intranet/apps/eighth/migrations/0062_auto_20200116_1926.py
 intranet/apps/eighth/migrations/0063_auto_20201224_1745.py
 intranet/apps/eighth/migrations/0064_auto_20210205_1153.py
 intranet/apps/eighth/migrations/__init__.py
+intranet/apps/eighth/tests/__init__.py
+intranet/apps/eighth/tests/test_admin.py
+intranet/apps/eighth/tests/test_general.py
+intranet/apps/eighth/tests/test_others.py
 intranet/apps/eighth/views/__init__.py
 intranet/apps/eighth/views/activities.py
 intranet/apps/eighth/views/api.py
