+++ docs/sourcedoc/intranet.apps.eighth.tests.rst
@@ -1,0 +1,37 @@
+intranet.apps.eighth.tests package
+==================================
+
+Submodules
+----------
+
+intranet.apps.eighth.tests.test\_admin module
+---------------------------------------------
+
+.. automodule:: intranet.apps.eighth.tests.test_admin
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+intranet.apps.eighth.tests.test\_general module
+-----------------------------------------------
+
+.. automodule:: intranet.apps.eighth.tests.test_general
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+intranet.apps.eighth.tests.test\_others module
+----------------------------------------------
+
+.. automodule:: intranet.apps.eighth.tests.test_others
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: intranet.apps.eighth.tests
+   :members:
+   :undoc-members:
+   :show-inheritance:
